move docker pruner to its own package
