+++ internal/cli/wire.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/google/wire"
 	"github.com/jonboulle/clockwork"
+	"github.com/windmilleng/tilt/internal/engine/dockerprune"
 	"github.com/windmilleng/wmclient/pkg/dirs"
 	"k8s.io/apimachinery/pkg/version"
 	"k8s.io/client-go/tools/clientcmd/api"
@@ -89,7 +90,7 @@ var BaseWireSet = wire.NewSet(
 	store.NewStore,
 	wire.Bind(new(store.RStore), new(store.Store)),
 
-	engine.NewDockerPruner,
+	dockerprune.NewDockerPruner,
 
 	provideTiltInfo,
 	engine.ProvideSubscribers,
