+++ internal/engine/upper_test.go
@@ -19,6 +19,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/windmilleng/tilt/internal/engine/dockerprune"
 	"github.com/windmilleng/wmclient/pkg/analytics"
 	"gopkg.in/yaml.v2"
 	v1 "k8s.io/api/core/v1"
@@ -2951,7 +2952,7 @@ func newTestFixture(t *testing.T) *testFixture {
 		tiltVersionCheckDelay: versionCheckInterval,
 	}
 
-	dp := NewDockerPruner(dockerClient)
+	dp := dockerprune.NewDockerPruner(dockerClient)
 	dp.DisableForTesting()
 
 	tiltVersionCheckTimerMaker := func(d time.Duration) <-chan time.Time {
